requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pgvector==0.3.3

# AI Framework
langchain==0.3.7
langgraph==0.2.40
langchain-openai==0.2.9
langchain-community==0.3.7
openai>=1.54.0,<2.0.0
anthropic==0.37.1
litellm==1.52.0
langgraph-checkpoint-postgres

# Feishu SDK
lark-oapi==1.4.6

# Config
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4

# Logging
loguru==0.7.2

# Document Processing
unstructured==0.15.14
python-docx==1.1.2

# RAG Evaluation
ragas==0.2.5

# Utils
requests==2.32.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Code Quality
ruff==0.7.0
